Added meetup information to a meetings page
Built off partials and linking directly to meetup.com the page includes
hData meant for easy computer parsing of Who, What, When, Where and Why
data. The index page now displays the upcoming event or a message about
events. Currently pages do not support images or further linking, but is
easy to add.